Initial docstrings and doctests for inference methods (#142)
* Initial doctest for `advancedHMC`

Simple docstring and test for the `advancedHMC` function. Notes:
  - Raw prefix to docstring is necessary to correctly parse the output filter.
  - E[β|x] is not deterministic, so it currently only checks the first two significant digits.

* finely control output generated by doctest

Switched from attempting to filter generated out, to switching off sampling progress printting and deprecation warnings.

* Change to flagging deprecation warnings

Revert the change to the travis config, and update the `advancedHMC` sampling call based on their current docs.

* initial docstring and doctest for `dynamicHMC`

This follows the docstring and doctest for `advancedHMC`. The `method` argument is undocumented because I couldn't figure out exactly what it is doing.

* Update src/inference/dynamicHMC.jl

Co-authored-by: Chad Scherrer <chad.scherrer@gmail.com>

Co-authored-by: Chad Scherrer <chad.scherrer@gmail.com>